The result of calling action is assigned to option
BACKWARD-INCOMPATIBLE change.  This change will break existing code
which depends on options with actions being dropped from
`*command-line-arguments*`.

Previously if an action was specified then the related option was
dropped from `*command-line-arguments*` causing any reference to it
from the main function to return nil.

With this change the return value of the function invoked by action is
now added to `*command-line-arguments*` and subsequently passed
through to the main function.