Bump version; incompatible change, new maintainer
Since f2063068 actions will modify the values of the corresponding
argument.  Specifically the value of the argument is set to the return
value of the defined action.  This is a BACKWARD INCOMPATIBLE change
requiring the bump in major version number.

This commit also updates the maintainer listed in the ASDF file.